requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pillow>=10.0.0
pytesseract>=0.3.10
pymupdf>=1.23.0
redis>=5.0.0
aiofiles>=23.2.0
